+++ src/main/kotlin/lsw/practice/domain/comment/model/Comment.kt
@@ -2,12 +2,12 @@ package lsw.practice.domain.comment.model
 
 import jakarta.persistence.*
 import lsw.practice.domain.post.model.Post
+import lsw.practice.domain.user.model.BaseTime
 import lsw.practice.domain.user.model.User
 import org.hibernate.annotations.OnDelete
 import org.hibernate.annotations.OnDeleteAction
 import org.hibernate.annotations.SQLDelete
 import org.hibernate.annotations.Where
-import java.time.LocalDateTime
 
 @Entity
 @SQLDelete(sql = "UPDATE comment SET is_deleted = true WHERE id = ?") // DELETE 쿼리 날아올 시 대신 실행
@@ -22,8 +22,8 @@ class Comment(
     @Column(name = "content", nullable = false)
     var content: String,
 
-    @Column(name = "created_at", nullable = false)
-    var createdAt: LocalDateTime = LocalDateTime.now(),
+//    @Column(name = "created_at", nullable = false)
+//    var createdAt: LocalDateTime = LocalDateTime.now(),
 
     @Column(name = "is_deleted", nullable = false)
     var isDeleted: Boolean = false,
@@ -35,7 +35,7 @@ class Comment(
     @ManyToOne()
     @JoinColumn(name = "post_id", nullable = false)
     var post: Post
-) {
+): BaseTime() {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null
