+++ src/main/kotlin/lsw/practice/domain/post/dto/PostResponse.kt
@@ -1,6 +1,7 @@
 package lsw.practice.domain.post.dto
 
 import lsw.practice.domain.comment.model.Comment
+import lsw.practice.domain.post.model.Post
 import java.time.LocalDateTime
 
 data class PostResponse(
@@ -8,6 +9,17 @@ data class PostResponse(
     val name: String,
     val title: String,
     val content: String,
-//    val createdAt: LocalDateTime,
     val comments: List<Comment>
-    )
+) {
+    companion object{
+        fun Post.toResponse(): PostResponse {
+            return PostResponse(
+                id = id!!,
+                title = title,
+                content = content,
+                name = name,
+                comments = comments
+            )
+        }
+    }
+}
