+++ src/main/kotlin/lsw/practice/domain/post/service/PostServiceImpl.kt
@@ -1,14 +1,13 @@
 package lsw.practice.domain.post.service
 
-import lsw.practice.domain.comment.repository.CommentRepository
 import lsw.practice.domain.exception.MismatchException
 import lsw.practice.domain.exception.ModelNotFoundException
 import lsw.practice.domain.post.dto.CreatePostRequest
 import lsw.practice.domain.post.dto.PostResponse
+import lsw.practice.domain.post.dto.PostResponse.Companion.toResponse
 import lsw.practice.domain.post.dto.UpdatePostRequest
 import lsw.practice.domain.post.model.Post
 import lsw.practice.domain.post.repository.PostRepository
-import lsw.practice.domain.post.repository.PostRepositoryImpl
 import lsw.practice.domain.user.repository.UserRepository
 import lsw.practice.infra.security.UserPrincipal
 import org.springframework.data.domain.Page
@@ -21,7 +20,7 @@ import org.springframework.transaction.annotation.Transactional
 class PostServiceImpl(
     private val postRepository: PostRepository,
     private val userRepository: UserRepository,
-): PostService {
+) : PostService {
     @Transactional
     override fun createPost(
         userPrincipal: UserPrincipal,
@@ -90,13 +89,3 @@ class PostServiceImpl(
     }
 }
 
-fun Post.toResponse(): PostResponse {
-    return PostResponse(
-        id = id!!,
-        title = title,
-        content = content,
-        name = name,
-//        createdAt = createdAt,
-        comments = comments
-    )
-}
