+++ src/main/kotlin/lsw/practice/domain/user/model/BaseTime.kt
@@ -1,0 +1,20 @@
+package lsw.practice.domain.user.model
+
+import jakarta.persistence.Column
+import jakarta.persistence.EntityListeners
+import jakarta.persistence.MappedSuperclass
+import org.hibernate.annotations.CreationTimestamp
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener::class)
+abstract class BaseTime {
+    @CreationTimestamp
+    @Column(nullable = false, updatable = false)
+    val createdAt: LocalDateTime? = null
+
+    @CreationTimestamp
+    @Column(nullable = false)
+    val updatedAt: LocalDateTime? = null
+}
